+++ blocks/coveo-search/coveo-search.js
@@ -1,9 +1,9 @@
 import { loadCSS } from '../../scripts/lib-franklin.js';
 import { loadScript, getCookie } from '../../scripts/scripts.js';
 
-const coveoAdminId = 'kapil.dhiman@moldev.com';
-const organizationId = 'moleculardevicesproductionca45f5xc';
-const coveoToken = 'xxb1966cc1-fee3-4df3-99bf-11fe2b4e1c74';
+export const coveoAdminId = 'kapil.dhiman@moldev.com';
+export const organizationId = 'moleculardevicesproductionca45f5xc';
+export const coveoToken = 'xxb1966cc1-fee3-4df3-99bf-11fe2b4e1c74';
 
 function getCategoriesBasedOnProfile(userProfile) {
   const CUSTOMER_ACCESS_LEVEL_CATEGORY = '"Customer"';
@@ -39,6 +39,7 @@ function getCategoriesBasedOnProfile(userProfile) {
 function getUserProfile() {
   return (getCookie('STYXKEY_PortalUserRole')) ? getCookie('STYXKEY_PortalUserRole') : '';
 }
+
 function getFilter() {
   const userProfile = getUserProfile();
   const accessLevel = getCategoriesBasedOnProfile(userProfile);
@@ -75,7 +76,7 @@ function searchFormHeader() {
         `;
 }
 
-function searchMainSection() {
+export function searchMainSection() {
   return `
             <div class="section coveo-tab-section-wrapper sticky-element sticky-desktop">
               <div class="coveo-tab-section">
@@ -94,7 +95,7 @@ function searchMainSection() {
                 <p class='coveoMainTitle coveo-main-title'></p>
                 <div class="coveo-facet-column">
                   <div class="CoveoDynamicFacet" data-enable-scroll-to-top="false" data-title="Country" data-field="@md_country" data-tab="SDS" data-id="Country" data-number-of-values="" data-enable-facet-search="false"></div>
-                  <div class="CoveoDynamicHierarchicalFacet" data-enable-facet-search="false" data-delimiting-character="|" data-title="Products" data-field="@mdproductsdatacategory" data-tab="Products, All, Resources, KBArticles, Videos" data-number-of-values="8" data-enable-collapse="true" data-enable-scroll-to-top="false" data-filter-facet-count="false"></div>
+                  <div class="CoveoDynamicHierarchicalFacet coveo-dynamic-hierarchical-facet" data-enable-facet-search="false" data-delimiting-character="|" data-title="Products" data-field="@mdproductsdatacategory" data-tab="Products, All, Resources, KBArticles, Videos" data-number-of-values="8" data-enable-collapse="true" data-enable-scroll-to-top="false" data-filter-facet-count="false"></div>
                   <div class="CoveoDynamicFacet" data-enable-scroll-to-top="false" data-title="Languages" data-field="@md_lang" data-tab="SDS" data-number-of-values="" data-depends-on="Country"></div>
                   <div class="CoveoDynamicFacet" data-enable-scroll-to-top="false" data-title="Type" data-field="@objecttype" data-tab="Resources"></div>
                   <div class="CoveoDynamicFacet" data-enable-scroll-to-top="false" data-title="Content Type" data-enable-facet-search="false" data-field="@md_contenttype" data-number-of-values="8" data-tab="Resources, All"></div>
@@ -221,7 +222,7 @@ async function coveoSearchInitiation(organizationID, accessToken) {
   });
 }
 
-async function getCoveoToken() {
+export async function getCoveoToken() {
   const myHeaders = new Headers();
   myHeaders.append('accept', 'application/json');
   myHeaders.append('Authorization', `Bearer ${coveoToken}`);
