+++ sitemap.xml
@@ -6,10 +6,7 @@
     <sitemap>
         <loc>https://www.moleculardevices.com/files-sitemap.xml</loc>
     </sitemap>
-<<<<<<< HEAD
     <sitemap>
         <loc>https://www.moleculardevices.com/hubspot-sitemap.xml</loc>
     </sitemap>
-=======
->>>>>>> 5951f26678ed0cca3af69b1cbf5013d7394d1784
 </sitemapindex>
